+++ GitHubClients/GitClient.ts
@@ -1,10 +1,10 @@
-import { GraphQlClient } from "../core/GraphQlClient.ts";
+import { GraphQlClient } from "../deps.ts";
 import { createGetBranchesQuery } from "../core/GraphQl/Queries/GetBranchesQuery.ts";
 import { Guard } from "../core/Guard.ts";
-import { PageInfoModel } from "../core/Models/mod.ts";
-import { GitBranchModel } from "../core/Models/mod.ts";
-import { RawRefsGetBranchModel } from "../core/Models/mod.ts";
-import { RawGitBranchModel } from "../core/Models/mod.ts";
+import { PageInfoModel } from "../deps.ts";
+import { GitBranchModel } from "../deps.ts";
+import { RawRefsGetBranchModel } from "../deps.ts";
+import { RawGitBranchModel } from "../deps.ts";
 import { Utils } from "../core/Utils.ts";
 import { RepoClient } from "./RepoClient.ts";
 import { getCreateBranchMutation } from "../core/GraphQl/Mutations/CreateBranchMutation.ts";
