+++ GitHubClients/IssueClient.ts
@@ -1,10 +1,10 @@
 import { Guard } from "../core/Guard.ts";
 import { LabelClient } from "./LabelClient.ts";
-import { IssueModel } from "../core/Models/mod.ts";
-import { LabelModel } from "../core/Models/mod.ts";
+import { IssueModel } from "../deps.ts";
+import { LabelModel } from "../deps.ts";
 import { Utils } from "../core/Utils.ts";
 import { GitHubHttpStatusCodes, IssueOrPRState } from "../core/Enums.ts";
-import { GitHubClient } from "../core/GitHubClient.ts";
+import { GitHubClient } from "../deps.ts";
 import { IssueOrPRRequestData } from "../core/IssueOrPRRequestData.ts";
 import { IssueError } from "./Errors/IssueError.ts";
