+++ GitHubClients/ProjectClient.ts
@@ -1,5 +1,5 @@
 import { createOrgProjectsQuery } from "../core/GraphQl/Queries/GetOrgProjectsQueries.ts";
-import { GraphQlClient } from "../core/GraphQlClient.ts";
+import { GraphQlClient } from "../deps.ts";
 import { Guard } from "../core/Guard.ts";
 import { Utils } from "../core/Utils.ts";
 import { createLinkItemToProjectMutation } from "../core/GraphQl/Mutations/AddToProjectMutation.ts";
@@ -7,9 +7,9 @@ import { createGetIssueProjectsQuery } from "../core/GraphQl/Queries/GetIssuePro
 import { createGetPullRequestProjectsQuery } from "../core/GraphQl/Queries/GetPullRequestProjectsQuery.ts";
 import { ProjectError } from "./Errors/ProjectError.ts";
 import { IssueClient, PullRequestClient } from "./mod.ts";
-import { ProjectModel } from "../core/Models/mod.ts";
-import { IssueModel } from "../core/Models/mod.ts";
-import { PullRequestModel } from "../core/Models/mod.ts";
+import { ProjectModel } from "../deps.ts";
+import { IssueModel } from "../deps.ts";
+import { PullRequestModel } from "../deps.ts";
 
 /**
  * Gets or saves data related to GitHub organization V2 projects.
