+++ GitHubClients/ReleaseClient.ts
@@ -1,7 +1,7 @@
 import { GitHubHttpStatusCodes } from "../core/Enums.ts";
-import { GitHubClient } from "../core/GitHubClient.ts";
+import { GitHubClient } from "../deps.ts";
 import { Guard } from "../core/Guard.ts";
-import { ReleaseModel } from "../core/Models/mod.ts";
+import { ReleaseModel } from "../deps.ts";
 import { Utils } from "../core/Utils.ts";
 import { basename, existsSync } from "../deps.ts";
 import { ReleaseError } from "./Errors/ReleaseError.ts";
