+++ GitHubClients/TagClient.ts
@@ -1,8 +1,8 @@
 import { Guard } from "../core/Guard.ts";
-import { TagModel } from "../core/Models/mod.ts";
+import { TagModel } from "../deps.ts";
 import { Utils } from "../core/Utils.ts";
 import { GitHubHttpStatusCodes } from "../core/Enums.ts";
-import { GitHubClient } from "../core/GitHubClient.ts";
+import { GitHubClient } from "../deps.ts";
 import { TagError } from "./Errors/TagError.ts";
 
 /**
