+++ GitHubClients/WorkflowClient.ts
@@ -1,9 +1,9 @@
 import { Guard } from "../core/Guard.ts";
 import { Utils } from "../core/Utils.ts";
 import { GitHubHttpStatusCodes, WorkflowEvent, WorkflowRunStatus } from "../core/Enums.ts";
-import { GitHubClient } from "../core/GitHubClient.ts";
-import { WorkflowRunModel } from "../core/Models/mod.ts";
-import { WorkflowRunsModel } from "../core/Models/mod.ts";
+import { GitHubClient } from "../deps.ts";
+import { WorkflowRunModel } from "../deps.ts";
+import { WorkflowRunsModel } from "../deps.ts";
 import { AnyBranch } from "../core/Types.ts";
 import { GithubResponse } from "../GitHubClients/GithubResponse.ts";
 import { WorkflowError } from "./Errors/WorkflowError.ts";
