+++ core/Models/GraphQlModels/GraphQlRequestResponseModel.ts
@@ -1,4 +1,4 @@
-import { ErrorModel } from "./ErrorModel.ts";
+import { ErrorModel } from "./mod.ts";
 
 /**
  * Represents a request response from a GraphQL request.
