+++ core/Models/GraphQlModels/RawModels/RawRefsGetBranchModel.ts
@@ -1,4 +1,4 @@
-import { RawGitBranchModel } from "./RawGitBranchModel.ts";
+import { RawGitBranchModel } from "./mod.ts";
 
 /**
  * Represents the raw git branch refs model that is unchanged from the GraphQL query.
